Gateway/external/real_ocr_service/requirements.txt
fastapi
uvicorn
python-multipart
onnxruntime
numpy
Pillow
